pyproject.toml
[project]
name = "desafio-01-engenharia"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicolas",email = "nicolascnagel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "apache-airflow (>=3.1.0,<4.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
